Extract public share pad links
