+++ js/etherpad_nextcloud-files-main.mjs
@@ -1,2 +1,2 @@
-import{a as L,A as f,o as oe,M as K,b as ye,c as $,d as we,e as be,n as he,f as ge}from"./oc-compat-DDPKm7Cs.chunk.mjs";import{g as j,n as k,p as ae,a as T,b as F,v as R,f as _,c as ve,r as se,i as q,d as Ce,e as Ee}from"./urls-B3dfoIpo.chunk.mjs";const S=new Map,H=async e=>{const i=String(e);if(S.has(i))return S.get(i);const n=L("/apps/"+f+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(i),r=I(n,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(a=>{throw S.delete(i),a});return S.set(i,r),r},xe=async e=>{const i="path:"+String(e);if(S.has(i))return S.get(i);const n=L("/apps/"+f+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),r=I(n,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(a=>{throw S.delete(i),a});return S.set(i,r),r},U=async e=>{const i=L("/apps/"+f+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e)));return I(i,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},Ae=async(e,i)=>{let n=L("/apps/"+f+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));return n+="?force=1",I(n,{method:"POST",headers:{Accept:"application/json",requesttoken:oe()}},"Pad sync failed.")},Ne=async(e,i)=>{const n=L("/apps/"+f+"/api/v1/pads/from-url"),r=new URLSearchParams;return r.set("file",e),r.set("padUrl",i),I(n,{method:"POST",headers:le(),body:r.toString()},"Could not import public pad URL.")},Se=async e=>{const i=L("/apps/"+f+"/api/v1/pads"),n=new URLSearchParams;return n.set("file",e),n.set("accessMode","public"),I(i,{method:"POST",headers:le(),body:n.toString()},"Could not create public pad.")},le=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:oe()}),I=async(e,i,n)=>{const r=await fetch(e,{...i,credentials:"same-origin"}),a=await r.json().catch(()=>({}));if(!r.ok)throw new Error(a&&a.message||n);return a},M=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),A=()=>(window.location.pathname||"").includes("/apps/files"),G="etherpad_nextcloud_open",D=e=>{const i=Number(e);return Number.isFinite(i)&&i>0?i:null},Le=e=>{if(!e)return null;const i=[e.fileId,e.fileid,e.id,e.nodeId,e.nodeid,e.file_id];for(const n of i){const r=D(n);if(r!==null)return r}if(typeof e.get=="function")for(const n of["fileid","fileId","id"])try{const r=D(e.get(n));if(r!==null)return r}catch{}return null},Pe=e=>{if(!(e instanceof Element))return null;const i=[e.getAttribute("data-fileid"),e.getAttribute("data-file-id"),e.getAttribute("data-id"),e.getAttribute("data-node-id")];for(const r of i){const a=D(r);if(a!==null)return a}const n=e.matches("a[href]")?e:e.querySelector("a[href]");return n instanceof HTMLAnchorElement?ae(n.getAttribute("href")):null},Me=e=>{if(!e||typeof e!="object")return null;const i=[e,e.fileInfoModel,e.fileInfo,e.model,e.node,e.file,e.attributes];for(const r of i){const a=Le(r);if(a!==null)return a}const n=[e.fileElement,e.el,e.element,e.$file&&e.$file[0],e.$el&&e.$el[0],e.target,e.currentTarget];for(const r of n){const a=Pe(r);if(a!==null)return a}return null},Ie=({openPadInNativeViewer:e})=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(K,G,ye(),$(f,"etherpad-icon-color"),(i,n)=>{const r=n&&n.dir||j(),a=k(r,i),s=Me(n);e({path:a,fileId:s})},t(f,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(K,G))},Oe=800,ke=180,Te=(e,i)=>{const n=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!n||typeof n.goToRoute!="function"||!e||!Number.isFinite(e)?!1:(n.goToRoute(null,{view:"files",fileid:String(e)},{dir:se(i),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(e)))return;const r=_(e,i);window.location.assign(r)},ke),!0)},Fe=()=>{let e=null,i=0;return async n=>{const r=String(n.fileId??"")+"|"+String(n.path??""),a=Date.now();if(e===r&&a-i<Oe)return;e=r,i=a;const s=T(),w=s!==null&&s!=="",c=()=>{if(w){window.location.assign(R(s,n.path||""));return}if(A()&&n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const C=n.path||"/";window.location.assign(_(Number(n.fileId),C));return}const h=A()?F():null;if((n.fileId===null||n.fileId===void 0)&&h){const C=n.path||"/";window.location.assign(_(h,C));return}if(n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const C=n.path||"/";window.location.assign(_(Number(n.fileId),C));return}n.path&&window.location.assign(ve(n.path))};if(!M()){c();return}let p=n.path||"",m=n.fileId??null;if(!p&&n.fileId!==null&&n.fileId!==void 0)try{const h=await H(n.fileId);p=h&&typeof h.path=="string"?h.path:"",m=h&&Number.isFinite(Number(h.file_id))?Number(h.file_id):m}catch{p=""}if(!p){c();return}if((!m||!Number.isFinite(Number(m)))&&p&&!w)try{const h=await xe(p);m=h&&Number.isFinite(Number(h.file_id))?Number(h.file_id):m}catch{}if((!m||!Number.isFinite(Number(m)))&&A()){const h=F();h&&(m=h)}if(A()){if(m&&Number.isFinite(Number(m))&&Te(Number(m),p))return;c();return}try{window.OCA.Viewer.open({path:p})}catch{c()}}},qe=120,Re=900,_e=async(e,i={})=>{if(M()&&e.path)try{const n=Number(e.fileId);A()&&Number.isFinite(n)&&n>0&&(Be(n,e.path,i.resolveOpenDir),await He()),await Ue(e.path),window.OCA.Viewer.open({path:e.path,onClose:De});return}catch{}typeof i.fallbackOpen=="function"&&await i.fallbackOpen(e)},He=()=>new Promise(e=>{window.setTimeout(e,qe)}),Ue=async e=>{const i=we(e);i!==""&&(be("editor:file:created",i),await ze())},ze=()=>new Promise(e=>{window.setTimeout(e,Re)}),Be=(e,i,n)=>{const r=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;!r||typeof r.goToRoute!="function"||r.goToRoute(null,{...r.params||{},view:r.params&&r.params.view?r.params.view:"files",fileid:String(e)},{...r.query||{},dir:je(i,n),editing:"false",openfile:"true"},!0)},je=(e,i)=>{if(typeof i=="function")return i(e);const n=String(e||""),r=n.lastIndexOf("/");return r>0?n.substring(0,r):"/"},De=()=>{const e=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;if(!e||typeof e.goToRoute!="function")return;const i={...e.query||{}};delete i.openfile,delete i.editing,e.goToRoute(null,e.params||{},i)},Ve=e=>{try{const i=new URL(e,window.location.origin),n=decodeURIComponent(i.pathname||"").match(/\/p\/([^/]+)$/);if(n&&n[1]){const r=n[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},ce=e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(f,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const a=document.createElement("h3");return a.textContent=e,a.style.margin="0 26px 10px 0",a.style.fontSize="18px",n.appendChild(r),n.appendChild(a),i.appendChild(n),document.body.appendChild(i),{overlay:i,dialog:n,closeButton:r}},$e=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=ce(t(f,"Internal public pad")),a=document.createElement("label");a.textContent=t(f,"File name"),a.style.display="block",a.style.marginBottom="6px";const s=document.createElement("input");s.type="text",s.value=t(f,"Public pad")+".pad",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const w=document.createElement("p");w.style.color="var(--color-error, #c62828)",w.style.margin="0 0 12px 0",w.style.minHeight="20px";const c=document.createElement("button");c.type="button",c.className="primary",c.textContent=t(f,"Create");const p=m=>{i.remove(),e(m)};r.addEventListener("click",()=>p(null)),i.addEventListener("click",m=>{m.target===i&&p(null)}),s.addEventListener("keydown",m=>{m.key==="Enter"&&(m.preventDefault(),c.click())}),c.addEventListener("click",()=>{const m=s.value.trim();if(m===""){w.textContent=t(f,"File name is required.");return}p(m)}),n.appendChild(a),n.appendChild(s),n.appendChild(w),n.appendChild(c),s.focus(),s.select()}),We=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=ce(t(f,"External public pad (URL)")),a=document.createElement("label");a.textContent=t(f,"External pad URL"),a.style.display="block",a.style.marginBottom="6px";const s=document.createElement("input");s.type="url",s.value="https://",s.placeholder="https://",s.style.width="100%",s.style.boxSizing="border-box",s.style.marginBottom="12px";const w=document.createElement("label");w.textContent=t(f,"File name"),w.style.display="block",w.style.marginBottom="6px";const c=document.createElement("input");c.type="text",c.value="Imported pad.pad",c.style.width="100%",c.style.boxSizing="border-box",c.style.marginBottom="12px";const p=document.createElement("p");p.style.color="var(--color-error, #c62828)",p.style.margin="0 0 12px 0",p.style.minHeight="20px";const m=document.createElement("button");m.type="button",m.className="primary",m.textContent=t(f,"Create");const h=g=>{i.remove(),e(g)};r.addEventListener("click",()=>h(null)),i.addEventListener("click",g=>{g.target===i&&h(null)}),s.addEventListener("blur",()=>{const g=s.value.trim();g.startsWith("http")&&(c.value=Ve(g))});const C=()=>{const g=s.value.trim(),u=c.value.trim();if(g===""){p.textContent=t(f,"External pad URL is required.");return}if(u===""){p.textContent=t(f,"File name is required.");return}h({padUrl:g,name:u})};s.addEventListener("keydown",g=>{g.key==="Enter"&&(g.preventDefault(),C())}),c.addEventListener("keydown",g=>{g.key==="Enter"&&(g.preventDefault(),C())}),m.addEventListener("click",C),n.appendChild(a),n.appendChild(s),n.appendChild(w),n.appendChild(c),n.appendChild(p),n.appendChild(m),s.focus(),s.select()}),Ke=()=>new Promise(e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(f,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const a=document.createElement("button");a.type="button",a.setAttribute("aria-label",t(f,"Close")),a.textContent="×",a.style.position="absolute",a.style.top="8px",a.style.right="10px",a.style.border="none",a.style.background="transparent",a.style.fontSize="22px",a.style.cursor="pointer",a.style.lineHeight="1";const s=document.createElement("p");s.textContent=t(f,"Choose how you want to create the public pad."),s.style.margin="0 0 14px 0",s.style.opacity="0.8";const w=document.createElement("div");w.style.display="grid",w.style.gap="8px";const c=document.createElement("button");c.type="button",c.className="primary",c.textContent=t(f,"Internal public pad");const p=document.createElement("button");p.type="button",p.textContent=t(f,"External public pad (URL)");const m=$(f,"etherpad-icon-color");m!==""&&(p.style.backgroundImage=`url(${m})`),p.style.backgroundRepeat="no-repeat",p.style.backgroundPosition="12px center",p.style.backgroundSize="16px 16px",p.style.paddingLeft="34px";const h=C=>{i.remove(),e(C)};a.addEventListener("click",()=>h(null)),c.addEventListener("click",()=>h("internal")),p.addEventListener("click",()=>h("external")),i.addEventListener("click",C=>{C.target===i&&h(null)}),w.appendChild(c),w.appendChild(p),n.appendChild(a),n.appendChild(r),n.appendChild(s),n.appendChild(w),i.appendChild(n),document.body.appendChild(i)}),J=e=>e.toLowerCase().endsWith(".pad")?e:e+".pad",Ge=(e,i)=>({path:e&&typeof e.file=="string"?e.file:i,fileId:e&&Number.isFinite(Number(e.file_id))?Number(e.file_id):null}),Je=({openPadInNativeViewer:e})=>{const i=(a,s)=>_e(Ge(a,s),{fallbackOpen:e,resolveOpenDir:se}),n=async()=>{const a=await $e();if(!a)return;const s=a.trim(),w=k(j(),J(s));try{const c=await Se(w);await i(c,w)}catch(c){const p=c instanceof Error?c.message:t(f,"Could not create public pad.");window.alert(p)}},r=async()=>{const a=await We();if(!a)return;const s=a.padUrl.trim(),w=a.name.trim(),c=k(j(),J(w));try{const p=await Ne(c,s);await i(p,c)}catch(p){const m=p instanceof Error?p.message:t(f,"Could not import public pad URL.");window.alert(m)}};return async()=>{const a=await Ke();if(a==="internal"){await n();return}a==="external"&&await r()}},z="data-epnc-sidebar-sync-panel",V="data-epnc-sidebar-sync-mount",de=8e3,Ze=6e4,ue=15e3,Qe=8e3,Xe=()=>{const e=document.documentElement,i=document.body,n=(e?.className||"")+" "+(i?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(n))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},Z=e=>{if(!(e instanceof HTMLElement))return!1;const i=window.getComputedStyle(e);return i.display==="none"||i.visibility==="hidden"?!1:i.position==="fixed"?!0:e.offsetParent!==null},Ye=e=>{const i=Number(e);return Number.isFinite(i)&&i>0?i:null},et=()=>{const e=F();if(e!==null)return e;const i=new URLSearchParams(window.location.search||""),n=Number(i.get("fileid")||"");return Number.isFinite(n)&&n>0?n:null},Q=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const i of e){const n=document.querySelector(i);if(n instanceof HTMLElement)return n}return null},tt=e=>{if(!(e instanceof HTMLElement))return null;let i=e.querySelector("["+V+'="1"]');i instanceof HTMLElement||(i=document.createElement("div"),i.className="epnc-sidebar-sync-mount",i.setAttribute(V,"1"));const n=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(n&&n.parentNode instanceof HTMLElement){const r=n.parentNode;return(i.parentNode!==r||i.nextSibling!==n)&&r.insertBefore(i,n),i}return i.parentNode!==e&&e.insertBefore(i,e.firstChild),i},X=e=>{const i=et();if(i!==null)return i;if(!(e instanceof HTMLElement))return null;const n=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const a of n){const s=e.querySelector(a);if(!(s instanceof HTMLElement))continue;const w=[s.getAttribute("data-fileid"),s.getAttribute("data-file-id"),s.getAttribute("data-node-id"),s.getAttribute("data-id"),s.dataset&&(s.dataset.fileid||s.dataset.fileId)];for(const c of w){const p=Ye(c);if(p!==null)return p}}const r=e.querySelectorAll("a[href]");for(const a of r){if(!(a instanceof HTMLAnchorElement))continue;const s=ae(a.getAttribute("href"));if(s!==null)return s}return null},nt=e=>e==="synced"?"etherpad-icon-color":Xe()?"etherpad-icon-white":"etherpad-icon-black",it=e=>e==="synced"?t(f,"Synchronized"):e==="syncing"?t(f,"Saving..."):e==="error"?t(f,"Sync failed"):e==="unavailable"?t(f,"Sync status unavailable"):e==="loading"?t(f,"Checking sync status..."):t(f,"Sync pending"),B=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",rt=(e,i)=>e==="synced"?Math.min(Ze,i+Qe):e==="pending"?de:ue,ot=()=>{let e=null,i=null,n=null,r=0;const a=()=>{n!==null&&(window.clearTimeout(n),n=null)},s=()=>{const u=document.querySelector("["+z+'="1"]');u&&u.parentNode&&u.parentNode.removeChild(u);const l=document.querySelector("["+V+'="1"]');l&&l.parentNode&&l.parentNode.removeChild(l),a()},w=(u,l)=>{if(!(u instanceof HTMLElement))return;const o=String(l||"").trim();u.setAttribute("data-public-open-url",o);const d=u.querySelector("[data-epnc-sidebar-open-public-button]");if(!(d instanceof HTMLButtonElement))return;const b=o!=="";d.style.display=b?"inline-flex":"none",d.disabled=!b},c=(u,l,o,d,b=!0)=>{if(!(u instanceof HTMLElement))return;const y=u.querySelector("[data-epnc-sidebar-sync-icon]"),E=u.querySelector("[data-epnc-sidebar-sync-label]"),v=u.querySelector("[data-epnc-sidebar-sync-button]");if(u.setAttribute("data-file-id",String(o)),u.setAttribute("data-sync-state",l),u.setAttribute("data-sync-in-flight",d?"1":"0"),E instanceof HTMLElement&&(E.textContent=it(l)),y instanceof HTMLElement){const N=nt(l),x=$(f,N);x!==""&&(y.style.backgroundImage=`url(${x})`)}v instanceof HTMLButtonElement&&(v.disabled=d||!b||o<=0,v.setAttribute("aria-busy",d?"true":"false"))},p=(u,l)=>{if(!(l instanceof HTMLElement))return null;let o=l.querySelector("["+z+'="1"]');if(!(o instanceof HTMLElement)){o=document.createElement("section"),o.className="epnc-sidebar-sync-panel",o.setAttribute(z,"1"),o.setAttribute("data-file-id",String(u));const d=document.createElement("div");d.className="epnc-sidebar-sync-panel__header";const b=document.createElement("span");b.className="epnc-sidebar-sync-panel__icon",b.setAttribute("data-epnc-sidebar-sync-icon","1"),b.setAttribute("aria-hidden","true");const y=document.createElement("span");y.className="epnc-sidebar-sync-panel__status",y.setAttribute("data-epnc-sidebar-sync-label","1"),d.appendChild(b),d.appendChild(y);const E=document.createElement("button");E.type="button",E.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",E.setAttribute("data-epnc-sidebar-sync-button","1"),E.textContent=t(f,"Pad in Datei speichern"),E.addEventListener("click",async N=>{N.preventDefault(),N.stopPropagation();const x=Number(o.getAttribute("data-file-id")||"");if(!(!Number.isFinite(x)||x<=0)){c(o,"syncing",x,!0,!0);try{await Ae(x,!0);const W=await U(x),me=B(W&&W.status);c(o,me,x,!1,!0)}catch{c(o,"error",x,!1,!0)}}});const v=document.createElement("button");v.type="button",v.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",v.setAttribute("data-epnc-sidebar-open-public-button","1"),v.textContent=t(f,"Pad in neuem Tab öffnen"),v.style.display="none",v.addEventListener("click",N=>{N.preventDefault(),N.stopPropagation();const x=String(o.getAttribute("data-public-open-url")||"").trim();x!==""&&window.open(x,"_blank","noopener,noreferrer")}),o.appendChild(d),o.appendChild(E),o.appendChild(v)}return o.parentNode!==l&&l.appendChild(o),o},m=(u,l)=>{if(a(),!(u instanceof HTMLElement)||!Number.isFinite(l)||l<=0)return;let o=de;const d=y=>{n=window.setTimeout(()=>{b()},Math.max(1e3,y))},b=async()=>{if(!(u instanceof HTMLElement)||!document.body.contains(u)){a();return}const y=Q();if(!(y instanceof HTMLElement)||!Z(y)||!Z(u)){a();return}if(X(y)!==l){a();return}if(document.visibilityState!=="visible"||u.getAttribute("data-sync-in-flight")==="1"){d(o);return}try{const E=await U(l),v=B(E&&E.status);c(u,v,l,!1,!0),o=rt(v,o)}catch{c(u,"error",l,!1,!0),o=ue}d(o)};d(o)},h=()=>{i!==null&&(window.clearTimeout(i),i=null)},C=async()=>{const u=++r;if(!A()||T()){s();return}const l=Q();if(!(l instanceof HTMLElement)){s();return}const o=X(l);if(!o){s();return}const d=tt(l);if(!(d instanceof HTMLElement)){s();return}let b;try{b=await H(o)}catch{s();return}if(u!==r)return;if(!b||b.is_pad!==!0){s();return}const y=p(o,d);if(!(y instanceof HTMLElement)){s();return}const E=b&&typeof b.public_open_url=="string"?b.public_open_url.trim():"";w(y,E),c(y,"loading",o,!1,!0);try{const v=await U(o);if(u!==r)return;const N=B(v&&v.status);c(y,N,o,!1,!0),m(y,o)}catch{if(u!==r)return;c(y,"error",o,!1,!0),m(y,o)}},g=(u=0)=>{h(),i=window.setTimeout(()=>{i=null,C()},Math.max(0,u))};return{installObserver:()=>{if(e!==null||!A())return;const u=new MutationObserver(l=>{for(const o of l)for(const d of o.addedNodes)if(d instanceof Element&&(d.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||d.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){g(60);return}});u.observe(document.body,{childList:!0,subtree:!0}),e=u},scheduleRefresh:g}},Y=f+"_public_pad",ee="icon-filetype-etherpad-nextcloud-pad",te='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',ne=98,at=120,st=500,ie=()=>{const e=he();return e!==null&&e>=33},O=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),lt=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const i=ge(),n=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const r of n){const a=Number(r);if(Number.isFinite(a))return(a&i)===i}return null},ct=(...e)=>{for(const i of e){const n=lt(i);if(n!==null)return n}return!0},dt=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const n of e)if(!(!n||typeof n!="object")){if(typeof n.addNewFileMenuEntry=="function")return{register:n.addNewFileMenuEntry.bind(n),unregister:typeof n.removeNewFileMenuEntry=="function"?n.removeNewFileMenuEntry.bind(n):null};if(typeof n.getNewFileMenu=="function"){let r=null;try{r=n.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const i=window._nc_newfilemenu;return i&&typeof i=="object"&&typeof i.registerEntry=="function"?{register:i.registerEntry.bind(i),unregister:typeof i.unregisterEntry=="function"?i.unregisterEntry.bind(i):null}:null},ut=({isFilesAppRoute:e,onCreatePublicPad:i})=>{let n=!1,r=!1,a=!1,s=0;const w=()=>{i()},c=(...g)=>ct(...g),p=()=>({id:Y,displayName:t(f,"Public pad"),...ie()?{iconSvgInline:te}:{iconClass:ee},order:ne,enabled:c,handler:w}),m=()=>({id:Y,displayName:t(f,"Public pad"),templateName:t(f,"Public pad"),iconClass:ee,...ie()?{iconSvgInline:te}:{},fileType:"file",order:ne,actionHandler:w}),h=async()=>{const g=dt();if(!g)return!1;const u=p();try{return g.register(u),n=!0,!0}catch(l){return O(l)?(n=!0,!0):!1}},C=()=>{if(r)return!0;const g=m(),u=()=>{const l=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const o of l)if(!(!o||typeof o!="object")){if(typeof o.addMenuEntry=="function")try{return o.addMenuEntry(g),!0}catch(d){if(O(d))return!0}if(typeof o.registerEntry=="function")try{return o.registerEntry(p()),!0}catch(d){if(O(d))return!0}}return!1};try{return u()?(r=!0,!0):(!a&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(l){if(!(!l||typeof l.addMenuEntry!="function"))try{l.addMenuEntry(g),r=!0}catch(o){O(o)&&(r=!0)}}}),a=!0),r)}catch(l){return O(l)?(r=!0,!0):!1}};return()=>{if(!e()||n||r)return;const g=++s,u=async l=>{g!==s||n||r||await h()||C()||l<at&&window.setTimeout(()=>{u(l+1)},st)};u(0)}},P="data-epnc-single-file-share";let re=0;const pe=e=>{if(!e)return null;const i='a[href*="/s/'+e+'/download"],a[href*="/index.php/s/'+e+'/download"]',n=document.querySelector(i);return n instanceof HTMLAnchorElement?n:null},pt=()=>{const e=document.getElementById("initial-state-files_sharing-view");if(!(e instanceof HTMLInputElement))return"";const i=String(e.value||"").trim();if(i==="")return"";try{const n=window.atob(i);if(n.startsWith('"')||n.startsWith("{")||n.startsWith("[")){const r=JSON.parse(n);return typeof r=="string"?r:""}return n}catch{return""}},fe=e=>{if(!e)return!1;const i=pe(e);if(!i)return!1;let n;try{n=new URL(i.getAttribute("href")||"",window.location.origin)}catch{return!1}const r=(n.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!r||r[1]!==e?!1:!n.searchParams.has("files")},ft=()=>{const e=document.body;if(!(e instanceof HTMLElement))return{retry:!0};const i=T();if(!i)return e.removeAttribute(P),{retry:!1};const n=pt();return n==="public-file-share"?(e.setAttribute(P,"1"),{retry:!1}):n==="public-share"?(e.removeAttribute(P),{retry:!1}):pe(i)?fe(i)?(e.setAttribute(P,"1"),{retry:!1}):(e.removeAttribute(P),{retry:!1}):(e.removeAttribute(P),{retry:!0})},mt=()=>{const e=++re,i=n=>{e===re&&ft().retry&&n<30&&window.setTimeout(()=>i(n+1),180)};i(0)};(function(){let e=!1,i="";const n=ot(),r=Fe(),a=Je({openPadInNativeViewer:r}),s=(l,o)=>{if(!(l instanceof HTMLAnchorElement)||!o)return null;const d=l.getAttribute("href")||"",b=Ce(d);if(b&&b.token===o&&q(b.path))return b.path;const y=Ee(d);return q(y)?y:null},w=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const l=async o=>{if(!o||o.defaultPrevented)return;const d=T();if(!d||o.type==="click"&&(o.button!==0||o.metaKey||o.ctrlKey||o.shiftKey||o.altKey)||!(o.target instanceof Element))return;const b=o.target.closest("a[href]");if(!(b instanceof HTMLAnchorElement))return;const y=s(b,d);if(y){if(o.preventDefault(),o.stopPropagation(),typeof o.stopImmediatePropagation=="function"&&o.stopImmediatePropagation(),M()){await r({path:y,fileId:null});return}window.location.assign(R(d,y))}};document.addEventListener("click",l,!0)},c=ut({isFilesAppRoute:A,onCreatePublicPad:a}),p=async()=>{if(M())return;const l=F();if(!l)return;const o=new URLSearchParams(window.location.search||"");if(!(o.get("epNoRedirect")==="1"||o.get("details")==="1"||o.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+f+"/"))try{const d=await H(l);d&&d.is_pad&&d.viewer_url&&window.location.assign(d.viewer_url)}catch{}},m=async()=>{if(M()||!A())return;const l=F();if(!l)return;const o=new URLSearchParams(window.location.search||"");if(!(o.get("openfile")==="true"||o.get("opendetails")==="true"||o.get("details")==="1"))try{const d=await H(l);if(!d||!d.is_pad)return;const b=o.get("dir")||"/",y=L("/apps/files/files")+"?dir="+encodeURIComponent(b);window.location.href!==y&&window.location.assign(y)}catch{}},h=()=>{const l=T();if(!l)return;const o=new URLSearchParams(window.location.search||""),d=o.get("files")||"";if(M()){if(!q(d))return;const E=o.get("path")||"/",v=k(E,d);r({path:v,fileId:null});return}if(!q(d)){if((window.location.pathname||"").includes("/apps/"+f+"/public/"))return;fe(l)&&window.location.assign(R(l,""));return}const b=o.get("path")||"/",y=k(b,d);window.location.assign(R(l,y))},C=()=>{const l=(window.location.pathname||"")+"?"+(window.location.search||"");l!==i&&(i=l,A()&&c(),mt(),m(),p(),h(),n.scheduleRefresh(0))},g=()=>{const l="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[l]===!0)return;window.OCA&&(window.OCA[l]=!0);const o=()=>window.setTimeout(C,0);window.addEventListener("popstate",o),window.addEventListener("hashchange",o);const d=b=>{const y=window.history&&window.history[b];typeof y=="function"&&(window.history[b]=function(...E){const v=y.apply(this,E);return o(),v})};d("pushState"),d("replaceState")},u=()=>{e||(e=!0,g(),C(),Ie({openPadInNativeViewer:r}),w(),n.installObserver(),n.scheduleRefresh(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",u,{once:!0}):u()})();
+import{a as L,A as f,o as oe,M as K,b as ye,c as $,d as we,e as be,n as he,f as ge}from"./oc-compat-DDPKm7Cs.chunk.mjs";import{g as j,n as k,p as ae,a as F,b as T,v as R,f as q,c as ve,r as se,d as Ce,i as _,e as Ee}from"./urls-B3dfoIpo.chunk.mjs";const S=new Map,H=async e=>{const i=String(e);if(S.has(i))return S.get(i);const n=L("/apps/"+f+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(i),r=M(n,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(o=>{throw S.delete(i),o});return S.set(i,r),r},xe=async e=>{const i="path:"+String(e);if(S.has(i))return S.get(i);const n=L("/apps/"+f+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),r=M(n,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(o=>{throw S.delete(i),o});return S.set(i,r),r},U=async e=>{const i=L("/apps/"+f+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e)));return M(i,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},Ae=async(e,i)=>{let n=L("/apps/"+f+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));return n+="?force=1",M(n,{method:"POST",headers:{Accept:"application/json",requesttoken:oe()}},"Pad sync failed.")},Ne=async(e,i)=>{const n=L("/apps/"+f+"/api/v1/pads/from-url"),r=new URLSearchParams;return r.set("file",e),r.set("padUrl",i),M(n,{method:"POST",headers:le(),body:r.toString()},"Could not import public pad URL.")},Se=async e=>{const i=L("/apps/"+f+"/api/v1/pads"),n=new URLSearchParams;return n.set("file",e),n.set("accessMode","public"),M(i,{method:"POST",headers:le(),body:n.toString()},"Could not create public pad.")},le=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:oe()}),M=async(e,i,n)=>{const r=await fetch(e,{...i,credentials:"same-origin"}),o=await r.json().catch(()=>({}));if(!r.ok)throw new Error(o&&o.message||n);return o},I=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),A=()=>(window.location.pathname||"").includes("/apps/files"),G="etherpad_nextcloud_open",D=e=>{const i=Number(e);return Number.isFinite(i)&&i>0?i:null},Le=e=>{if(!e)return null;const i=[e.fileId,e.fileid,e.id,e.nodeId,e.nodeid,e.file_id];for(const n of i){const r=D(n);if(r!==null)return r}if(typeof e.get=="function")for(const n of["fileid","fileId","id"])try{const r=D(e.get(n));if(r!==null)return r}catch{}return null},Pe=e=>{if(!(e instanceof Element))return null;const i=[e.getAttribute("data-fileid"),e.getAttribute("data-file-id"),e.getAttribute("data-id"),e.getAttribute("data-node-id")];for(const r of i){const o=D(r);if(o!==null)return o}const n=e.matches("a[href]")?e:e.querySelector("a[href]");return n instanceof HTMLAnchorElement?ae(n.getAttribute("href")):null},Ie=e=>{if(!e||typeof e!="object")return null;const i=[e,e.fileInfoModel,e.fileInfo,e.model,e.node,e.file,e.attributes];for(const r of i){const o=Le(r);if(o!==null)return o}const n=[e.fileElement,e.el,e.element,e.$file&&e.$file[0],e.$el&&e.$el[0],e.target,e.currentTarget];for(const r of n){const o=Pe(r);if(o!==null)return o}return null},Me=({openPadInNativeViewer:e})=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(K,G,ye(),$(f,"etherpad-icon-color"),(i,n)=>{const r=n&&n.dir||j(),o=k(r,i),a=Ie(n);e({path:o,fileId:a})},t(f,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(K,G))},Oe=800,ke=180,Te=(e,i)=>{const n=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!n||typeof n.goToRoute!="function"||!e||!Number.isFinite(e)?!1:(n.goToRoute(null,{view:"files",fileid:String(e)},{dir:se(i),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(e)))return;const r=q(e,i);window.location.assign(r)},ke),!0)},Fe=()=>{let e=null,i=0;return async n=>{const r=String(n.fileId??"")+"|"+String(n.path??""),o=Date.now();if(e===r&&o-i<Oe)return;e=r,i=o;const a=F(),w=a!==null&&a!=="",d=()=>{if(w){window.location.assign(R(a,n.path||""));return}if(A()&&n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const v=n.path||"/";window.location.assign(q(Number(n.fileId),v));return}const b=A()?T():null;if((n.fileId===null||n.fileId===void 0)&&b){const v=n.path||"/";window.location.assign(q(b,v));return}if(n.fileId!==null&&n.fileId!==void 0&&Number.isFinite(Number(n.fileId))){const v=n.path||"/";window.location.assign(q(Number(n.fileId),v));return}n.path&&window.location.assign(ve(n.path))};if(!I()){d();return}let u=n.path||"",m=n.fileId??null;if(!u&&n.fileId!==null&&n.fileId!==void 0)try{const b=await H(n.fileId);u=b&&typeof b.path=="string"?b.path:"",m=b&&Number.isFinite(Number(b.file_id))?Number(b.file_id):m}catch{u=""}if(!u){d();return}if((!m||!Number.isFinite(Number(m)))&&u&&!w)try{const b=await xe(u);m=b&&Number.isFinite(Number(b.file_id))?Number(b.file_id):m}catch{}if((!m||!Number.isFinite(Number(m)))&&A()){const b=T();b&&(m=b)}if(A()){if(m&&Number.isFinite(Number(m))&&Te(Number(m),u))return;d();return}try{window.OCA.Viewer.open({path:u})}catch{d()}}},qe=120,Re=900,_e=async(e,i={})=>{if(I()&&e.path)try{const n=Number(e.fileId);A()&&Number.isFinite(n)&&n>0&&(Be(n,e.path,i.resolveOpenDir),await He()),await Ue(e.path),window.OCA.Viewer.open({path:e.path,onClose:De});return}catch{}typeof i.fallbackOpen=="function"&&await i.fallbackOpen(e)},He=()=>new Promise(e=>{window.setTimeout(e,qe)}),Ue=async e=>{const i=we(e);i!==""&&(be("editor:file:created",i),await ze())},ze=()=>new Promise(e=>{window.setTimeout(e,Re)}),Be=(e,i,n)=>{const r=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;!r||typeof r.goToRoute!="function"||r.goToRoute(null,{...r.params||{},view:r.params&&r.params.view?r.params.view:"files",fileid:String(e)},{...r.query||{},dir:je(i,n),editing:"false",openfile:"true"},!0)},je=(e,i)=>{if(typeof i=="function")return i(e);const n=String(e||""),r=n.lastIndexOf("/");return r>0?n.substring(0,r):"/"},De=()=>{const e=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;if(!e||typeof e.goToRoute!="function")return;const i={...e.query||{}};delete i.openfile,delete i.editing,e.goToRoute(null,e.params||{},i)},Ve=e=>{try{const i=new URL(e,window.location.origin),n=decodeURIComponent(i.pathname||"").match(/\/p\/([^/]+)$/);if(n&&n[1]){const r=n[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(r!=="")return r+".pad"}}catch{}return"Imported pad.pad"},ce=e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(460px, calc(100vw - 24px))";const r=document.createElement("button");r.type="button",r.setAttribute("aria-label",t(f,"Close")),r.textContent="×",r.style.position="absolute",r.style.top="8px",r.style.right="10px",r.style.border="none",r.style.background="transparent",r.style.fontSize="22px",r.style.cursor="pointer",r.style.lineHeight="1";const o=document.createElement("h3");return o.textContent=e,o.style.margin="0 26px 10px 0",o.style.fontSize="18px",n.appendChild(r),n.appendChild(o),i.appendChild(n),document.body.appendChild(i),{overlay:i,dialog:n,closeButton:r}},$e=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=ce(t(f,"Internal public pad")),o=document.createElement("label");o.textContent=t(f,"File name"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="text",a.value=t(f,"Public pad")+".pad",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const w=document.createElement("p");w.style.color="var(--color-error, #c62828)",w.style.margin="0 0 12px 0",w.style.minHeight="20px";const d=document.createElement("button");d.type="button",d.className="primary",d.textContent=t(f,"Create");const u=m=>{i.remove(),e(m)};r.addEventListener("click",()=>u(null)),i.addEventListener("click",m=>{m.target===i&&u(null)}),a.addEventListener("keydown",m=>{m.key==="Enter"&&(m.preventDefault(),d.click())}),d.addEventListener("click",()=>{const m=a.value.trim();if(m===""){w.textContent=t(f,"File name is required.");return}u(m)}),n.appendChild(o),n.appendChild(a),n.appendChild(w),n.appendChild(d),a.focus(),a.select()}),We=()=>new Promise(e=>{const{overlay:i,dialog:n,closeButton:r}=ce(t(f,"External public pad (URL)")),o=document.createElement("label");o.textContent=t(f,"External pad URL"),o.style.display="block",o.style.marginBottom="6px";const a=document.createElement("input");a.type="url",a.value="https://",a.placeholder="https://",a.style.width="100%",a.style.boxSizing="border-box",a.style.marginBottom="12px";const w=document.createElement("label");w.textContent=t(f,"File name"),w.style.display="block",w.style.marginBottom="6px";const d=document.createElement("input");d.type="text",d.value="Imported pad.pad",d.style.width="100%",d.style.boxSizing="border-box",d.style.marginBottom="12px";const u=document.createElement("p");u.style.color="var(--color-error, #c62828)",u.style.margin="0 0 12px 0",u.style.minHeight="20px";const m=document.createElement("button");m.type="button",m.className="primary",m.textContent=t(f,"Create");const b=p=>{i.remove(),e(p)};r.addEventListener("click",()=>b(null)),i.addEventListener("click",p=>{p.target===i&&b(null)}),a.addEventListener("blur",()=>{const p=a.value.trim();p.startsWith("http")&&(d.value=Ve(p))});const v=()=>{const p=a.value.trim(),s=d.value.trim();if(p===""){u.textContent=t(f,"External pad URL is required.");return}if(s===""){u.textContent=t(f,"File name is required.");return}b({padUrl:p,name:s})};a.addEventListener("keydown",p=>{p.key==="Enter"&&(p.preventDefault(),v())}),d.addEventListener("keydown",p=>{p.key==="Enter"&&(p.preventDefault(),v())}),m.addEventListener("click",v),n.appendChild(o),n.appendChild(a),n.appendChild(w),n.appendChild(d),n.appendChild(u),n.appendChild(m),a.focus(),a.select()}),Ke=()=>new Promise(e=>{const i=document.createElement("div");i.style.position="fixed",i.style.inset="0",i.style.background="rgba(0, 0, 0, 0.45)",i.style.display="flex",i.style.alignItems="center",i.style.justifyContent="center",i.style.zIndex="20000";const n=document.createElement("div");n.style.position="relative",n.style.background="var(--color-main-background, #fff)",n.style.borderRadius="10px",n.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",n.style.padding="18px",n.style.width="min(420px, calc(100vw - 24px))";const r=document.createElement("h3");r.textContent=t(f,"Create public pad"),r.style.margin="0 26px 10px 0",r.style.fontSize="18px";const o=document.createElement("button");o.type="button",o.setAttribute("aria-label",t(f,"Close")),o.textContent="×",o.style.position="absolute",o.style.top="8px",o.style.right="10px",o.style.border="none",o.style.background="transparent",o.style.fontSize="22px",o.style.cursor="pointer",o.style.lineHeight="1";const a=document.createElement("p");a.textContent=t(f,"Choose how you want to create the public pad."),a.style.margin="0 0 14px 0",a.style.opacity="0.8";const w=document.createElement("div");w.style.display="grid",w.style.gap="8px";const d=document.createElement("button");d.type="button",d.className="primary",d.textContent=t(f,"Internal public pad");const u=document.createElement("button");u.type="button",u.textContent=t(f,"External public pad (URL)");const m=$(f,"etherpad-icon-color");m!==""&&(u.style.backgroundImage=`url(${m})`),u.style.backgroundRepeat="no-repeat",u.style.backgroundPosition="12px center",u.style.backgroundSize="16px 16px",u.style.paddingLeft="34px";const b=v=>{i.remove(),e(v)};o.addEventListener("click",()=>b(null)),d.addEventListener("click",()=>b("internal")),u.addEventListener("click",()=>b("external")),i.addEventListener("click",v=>{v.target===i&&b(null)}),w.appendChild(d),w.appendChild(u),n.appendChild(o),n.appendChild(r),n.appendChild(a),n.appendChild(w),i.appendChild(n),document.body.appendChild(i)}),J=e=>e.toLowerCase().endsWith(".pad")?e:e+".pad",Ge=(e,i)=>({path:e&&typeof e.file=="string"?e.file:i,fileId:e&&Number.isFinite(Number(e.file_id))?Number(e.file_id):null}),Je=({openPadInNativeViewer:e})=>{const i=(o,a)=>_e(Ge(o,a),{fallbackOpen:e,resolveOpenDir:se}),n=async()=>{const o=await $e();if(!o)return;const a=o.trim(),w=k(j(),J(a));try{const d=await Se(w);await i(d,w)}catch(d){const u=d instanceof Error?d.message:t(f,"Could not create public pad.");window.alert(u)}},r=async()=>{const o=await We();if(!o)return;const a=o.padUrl.trim(),w=o.name.trim(),d=k(j(),J(w));try{const u=await Ne(d,a);await i(u,d)}catch(u){const m=u instanceof Error?u.message:t(f,"Could not import public pad URL.");window.alert(m)}};return async()=>{const o=await Ke();if(o==="internal"){await n();return}o==="external"&&await r()}},z="data-epnc-sidebar-sync-panel",V="data-epnc-sidebar-sync-mount",de=8e3,Ze=6e4,ue=15e3,Qe=8e3,Xe=()=>{const e=document.documentElement,i=document.body,n=(e?.className||"")+" "+(i?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(n))return!0;const r=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(r&&r.matches)},Z=e=>{if(!(e instanceof HTMLElement))return!1;const i=window.getComputedStyle(e);return i.display==="none"||i.visibility==="hidden"?!1:i.position==="fixed"?!0:e.offsetParent!==null},Ye=e=>{const i=Number(e);return Number.isFinite(i)&&i>0?i:null},et=()=>{const e=T();if(e!==null)return e;const i=new URLSearchParams(window.location.search||""),n=Number(i.get("fileid")||"");return Number.isFinite(n)&&n>0?n:null},Q=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const i of e){const n=document.querySelector(i);if(n instanceof HTMLElement)return n}return null},tt=e=>{if(!(e instanceof HTMLElement))return null;let i=e.querySelector("["+V+'="1"]');i instanceof HTMLElement||(i=document.createElement("div"),i.className="epnc-sidebar-sync-mount",i.setAttribute(V,"1"));const n=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(n&&n.parentNode instanceof HTMLElement){const r=n.parentNode;return(i.parentNode!==r||i.nextSibling!==n)&&r.insertBefore(i,n),i}return i.parentNode!==e&&e.insertBefore(i,e.firstChild),i},X=e=>{const i=et();if(i!==null)return i;if(!(e instanceof HTMLElement))return null;const n=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const o of n){const a=e.querySelector(o);if(!(a instanceof HTMLElement))continue;const w=[a.getAttribute("data-fileid"),a.getAttribute("data-file-id"),a.getAttribute("data-node-id"),a.getAttribute("data-id"),a.dataset&&(a.dataset.fileid||a.dataset.fileId)];for(const d of w){const u=Ye(d);if(u!==null)return u}}const r=e.querySelectorAll("a[href]");for(const o of r){if(!(o instanceof HTMLAnchorElement))continue;const a=ae(o.getAttribute("href"));if(a!==null)return a}return null},nt=e=>e==="synced"?"etherpad-icon-color":Xe()?"etherpad-icon-white":"etherpad-icon-black",it=e=>e==="synced"?t(f,"Synchronized"):e==="syncing"?t(f,"Saving..."):e==="error"?t(f,"Sync failed"):e==="unavailable"?t(f,"Sync status unavailable"):e==="loading"?t(f,"Checking sync status..."):t(f,"Sync pending"),B=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",rt=(e,i)=>e==="synced"?Math.min(Ze,i+Qe):e==="pending"?de:ue,ot=()=>{let e=null,i=null,n=null,r=0;const o=()=>{n!==null&&(window.clearTimeout(n),n=null)},a=()=>{const s=document.querySelector("["+z+'="1"]');s&&s.parentNode&&s.parentNode.removeChild(s);const l=document.querySelector("["+V+'="1"]');l&&l.parentNode&&l.parentNode.removeChild(l),o()},w=(s,l)=>{if(!(s instanceof HTMLElement))return;const c=String(l||"").trim();s.setAttribute("data-public-open-url",c);const y=s.querySelector("[data-epnc-sidebar-open-public-button]");if(!(y instanceof HTMLButtonElement))return;const g=c!=="";y.style.display=g?"inline-flex":"none",y.disabled=!g},d=(s,l,c,y,g=!0)=>{if(!(s instanceof HTMLElement))return;const h=s.querySelector("[data-epnc-sidebar-sync-icon]"),E=s.querySelector("[data-epnc-sidebar-sync-label]"),C=s.querySelector("[data-epnc-sidebar-sync-button]");if(s.setAttribute("data-file-id",String(c)),s.setAttribute("data-sync-state",l),s.setAttribute("data-sync-in-flight",y?"1":"0"),E instanceof HTMLElement&&(E.textContent=it(l)),h instanceof HTMLElement){const N=nt(l),x=$(f,N);x!==""&&(h.style.backgroundImage=`url(${x})`)}C instanceof HTMLButtonElement&&(C.disabled=y||!g||c<=0,C.setAttribute("aria-busy",y?"true":"false"))},u=(s,l)=>{if(!(l instanceof HTMLElement))return null;let c=l.querySelector("["+z+'="1"]');if(!(c instanceof HTMLElement)){c=document.createElement("section"),c.className="epnc-sidebar-sync-panel",c.setAttribute(z,"1"),c.setAttribute("data-file-id",String(s));const y=document.createElement("div");y.className="epnc-sidebar-sync-panel__header";const g=document.createElement("span");g.className="epnc-sidebar-sync-panel__icon",g.setAttribute("data-epnc-sidebar-sync-icon","1"),g.setAttribute("aria-hidden","true");const h=document.createElement("span");h.className="epnc-sidebar-sync-panel__status",h.setAttribute("data-epnc-sidebar-sync-label","1"),y.appendChild(g),y.appendChild(h);const E=document.createElement("button");E.type="button",E.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",E.setAttribute("data-epnc-sidebar-sync-button","1"),E.textContent=t(f,"Pad in Datei speichern"),E.addEventListener("click",async N=>{N.preventDefault(),N.stopPropagation();const x=Number(c.getAttribute("data-file-id")||"");if(!(!Number.isFinite(x)||x<=0)){d(c,"syncing",x,!0,!0);try{await Ae(x,!0);const W=await U(x),me=B(W&&W.status);d(c,me,x,!1,!0)}catch{d(c,"error",x,!1,!0)}}});const C=document.createElement("button");C.type="button",C.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",C.setAttribute("data-epnc-sidebar-open-public-button","1"),C.textContent=t(f,"Pad in neuem Tab öffnen"),C.style.display="none",C.addEventListener("click",N=>{N.preventDefault(),N.stopPropagation();const x=String(c.getAttribute("data-public-open-url")||"").trim();x!==""&&window.open(x,"_blank","noopener,noreferrer")}),c.appendChild(y),c.appendChild(E),c.appendChild(C)}return c.parentNode!==l&&l.appendChild(c),c},m=(s,l)=>{if(o(),!(s instanceof HTMLElement)||!Number.isFinite(l)||l<=0)return;let c=de;const y=h=>{n=window.setTimeout(()=>{g()},Math.max(1e3,h))},g=async()=>{if(!(s instanceof HTMLElement)||!document.body.contains(s)){o();return}const h=Q();if(!(h instanceof HTMLElement)||!Z(h)||!Z(s)){o();return}if(X(h)!==l){o();return}if(document.visibilityState!=="visible"||s.getAttribute("data-sync-in-flight")==="1"){y(c);return}try{const E=await U(l),C=B(E&&E.status);d(s,C,l,!1,!0),c=rt(C,c)}catch{d(s,"error",l,!1,!0),c=ue}y(c)};y(c)},b=()=>{i!==null&&(window.clearTimeout(i),i=null)},v=async()=>{const s=++r;if(!A()||F()){a();return}const l=Q();if(!(l instanceof HTMLElement)){a();return}const c=X(l);if(!c){a();return}const y=tt(l);if(!(y instanceof HTMLElement)){a();return}let g;try{g=await H(c)}catch{a();return}if(s!==r)return;if(!g||g.is_pad!==!0){a();return}const h=u(c,y);if(!(h instanceof HTMLElement)){a();return}const E=g&&typeof g.public_open_url=="string"?g.public_open_url.trim():"";w(h,E),d(h,"loading",c,!1,!0);try{const C=await U(c);if(s!==r)return;const N=B(C&&C.status);d(h,N,c,!1,!0),m(h,c)}catch{if(s!==r)return;d(h,"error",c,!1,!0),m(h,c)}},p=(s=0)=>{b(),i=window.setTimeout(()=>{i=null,v()},Math.max(0,s))};return{installObserver:()=>{if(e!==null||!A())return;const s=new MutationObserver(l=>{for(const c of l)for(const y of c.addedNodes)if(y instanceof Element&&(y.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||y.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){p(60);return}});s.observe(document.body,{childList:!0,subtree:!0}),e=s},scheduleRefresh:p}},Y=f+"_public_pad",ee="icon-filetype-etherpad-nextcloud-pad",te='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',ne=98,at=120,st=500,ie=()=>{const e=he();return e!==null&&e>=33},O=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),lt=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const i=ge(),n=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const r of n){const o=Number(r);if(Number.isFinite(o))return(o&i)===i}return null},ct=(...e)=>{for(const i of e){const n=lt(i);if(n!==null)return n}return!0},dt=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const n of e)if(!(!n||typeof n!="object")){if(typeof n.addNewFileMenuEntry=="function")return{register:n.addNewFileMenuEntry.bind(n),unregister:typeof n.removeNewFileMenuEntry=="function"?n.removeNewFileMenuEntry.bind(n):null};if(typeof n.getNewFileMenu=="function"){let r=null;try{r=n.getNewFileMenu()}catch{r=null}if(r&&typeof r.registerEntry=="function")return{register:r.registerEntry.bind(r),unregister:typeof r.unregisterEntry=="function"?r.unregisterEntry.bind(r):null}}}const i=window._nc_newfilemenu;return i&&typeof i=="object"&&typeof i.registerEntry=="function"?{register:i.registerEntry.bind(i),unregister:typeof i.unregisterEntry=="function"?i.unregisterEntry.bind(i):null}:null},ut=({isFilesAppRoute:e,onCreatePublicPad:i})=>{let n=!1,r=!1,o=!1,a=0;const w=()=>{i()},d=(...p)=>ct(...p),u=()=>({id:Y,displayName:t(f,"Public pad"),...ie()?{iconSvgInline:te}:{iconClass:ee},order:ne,enabled:d,handler:w}),m=()=>({id:Y,displayName:t(f,"Public pad"),templateName:t(f,"Public pad"),iconClass:ee,...ie()?{iconSvgInline:te}:{},fileType:"file",order:ne,actionHandler:w}),b=async()=>{const p=dt();if(!p)return!1;const s=u();try{return p.register(s),n=!0,!0}catch(l){return O(l)?(n=!0,!0):!1}},v=()=>{if(r)return!0;const p=m(),s=()=>{const l=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const c of l)if(!(!c||typeof c!="object")){if(typeof c.addMenuEntry=="function")try{return c.addMenuEntry(p),!0}catch(y){if(O(y))return!0}if(typeof c.registerEntry=="function")try{return c.registerEntry(u()),!0}catch(y){if(O(y))return!0}}return!1};try{return s()?(r=!0,!0):(!o&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(l){if(!(!l||typeof l.addMenuEntry!="function"))try{l.addMenuEntry(p),r=!0}catch(c){O(c)&&(r=!0)}}}),o=!0),r)}catch(l){return O(l)?(r=!0,!0):!1}};return()=>{if(!e()||n||r)return;const p=++a,s=async l=>{p!==a||n||r||await b()||v()||l<at&&window.setTimeout(()=>{s(l+1)},st)};s(0)}},pt=(e,i)=>{if(!(e instanceof HTMLAnchorElement)||!i)return null;const n=e.getAttribute("href")||"",r=Ce(n);if(r&&r.token===i&&_(r.path))return r.path;const o=Ee(n);return _(o)?o:null},ft=({openPadInNativeViewer:e})=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const i=async n=>{if(!n||n.defaultPrevented)return;const r=F();if(!r||n.type==="click"&&(n.button!==0||n.metaKey||n.ctrlKey||n.shiftKey||n.altKey)||!(n.target instanceof Element))return;const o=n.target.closest("a[href]");if(!(o instanceof HTMLAnchorElement))return;const a=pt(o,r);if(a){if(n.preventDefault(),n.stopPropagation(),typeof n.stopImmediatePropagation=="function"&&n.stopImmediatePropagation(),I()){await e({path:a,fileId:null});return}window.location.assign(R(r,a))}};document.addEventListener("click",i,!0)},P="data-epnc-single-file-share";let re=0;const pe=e=>{if(!e)return null;const i='a[href*="/s/'+e+'/download"],a[href*="/index.php/s/'+e+'/download"]',n=document.querySelector(i);return n instanceof HTMLAnchorElement?n:null},mt=()=>{const e=document.getElementById("initial-state-files_sharing-view");if(!(e instanceof HTMLInputElement))return"";const i=String(e.value||"").trim();if(i==="")return"";try{const n=window.atob(i);if(n.startsWith('"')||n.startsWith("{")||n.startsWith("[")){const r=JSON.parse(n);return typeof r=="string"?r:""}return n}catch{return""}},fe=e=>{if(!e)return!1;const i=pe(e);if(!i)return!1;let n;try{n=new URL(i.getAttribute("href")||"",window.location.origin)}catch{return!1}const r=(n.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!r||r[1]!==e?!1:!n.searchParams.has("files")},yt=()=>{const e=document.body;if(!(e instanceof HTMLElement))return{retry:!0};const i=F();if(!i)return e.removeAttribute(P),{retry:!1};const n=mt();return n==="public-file-share"?(e.setAttribute(P,"1"),{retry:!1}):n==="public-share"?(e.removeAttribute(P),{retry:!1}):pe(i)?fe(i)?(e.setAttribute(P,"1"),{retry:!1}):(e.removeAttribute(P),{retry:!1}):(e.removeAttribute(P),{retry:!0})},wt=()=>{const e=++re,i=n=>{e===re&&yt().retry&&n<30&&window.setTimeout(()=>i(n+1),180)};i(0)};(function(){let e=!1,i="";const n=ot(),r=Fe(),o=Je({openPadInNativeViewer:r}),a=ut({isFilesAppRoute:A,onCreatePublicPad:o}),w=async()=>{if(I())return;const p=T();if(!p)return;const s=new URLSearchParams(window.location.search||"");if(!(s.get("epNoRedirect")==="1"||s.get("details")==="1"||s.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+f+"/"))try{const l=await H(p);l&&l.is_pad&&l.viewer_url&&window.location.assign(l.viewer_url)}catch{}},d=async()=>{if(I()||!A())return;const p=T();if(!p)return;const s=new URLSearchParams(window.location.search||"");if(!(s.get("openfile")==="true"||s.get("opendetails")==="true"||s.get("details")==="1"))try{const l=await H(p);if(!l||!l.is_pad)return;const c=s.get("dir")||"/",y=L("/apps/files/files")+"?dir="+encodeURIComponent(c);window.location.href!==y&&window.location.assign(y)}catch{}},u=()=>{const p=F();if(!p)return;const s=new URLSearchParams(window.location.search||""),l=s.get("files")||"";if(I()){if(!_(l))return;const g=s.get("path")||"/",h=k(g,l);r({path:h,fileId:null});return}if(!_(l)){if((window.location.pathname||"").includes("/apps/"+f+"/public/"))return;fe(p)&&window.location.assign(R(p,""));return}const c=s.get("path")||"/",y=k(c,l);window.location.assign(R(p,y))},m=()=>{const p=(window.location.pathname||"")+"?"+(window.location.search||"");p!==i&&(i=p,A()&&a(),wt(),d(),w(),u(),n.scheduleRefresh(0))},b=()=>{const p="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[p]===!0)return;window.OCA&&(window.OCA[p]=!0);const s=()=>window.setTimeout(m,0);window.addEventListener("popstate",s),window.addEventListener("hashchange",s);const l=c=>{const y=window.history&&window.history[c];typeof y=="function"&&(window.history[c]=function(...g){const h=y.apply(this,g);return s(),h})};l("pushState"),l("replaceState")},v=()=>{e||(e=!0,b(),m(),Me({openPadInNativeViewer:r}),ft({openPadInNativeViewer:r}),n.installObserver(),n.scheduleRefresh(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",v,{once:!0}):v()})();
 //# sourceMappingURL=etherpad_nextcloud-files-main.mjs.map
